requirements.txt
numpy
pandas
scipy
matplotlib
seaborn
scikit-learn
scikit-image
pillow
opencv-python
tensorflow
keras
jupyter